move ConfigurationYaml class to internal.deployment.parser package
